+++ backend/app/rest/proxy/image.go
@@ -6,7 +6,6 @@ import (
 	"encoding/base64"
 	"fmt"
 	"io"
-	"net"
 	"net/http"
 	"strings"
 	"time"
@@ -16,6 +15,7 @@ import (
 	"github.com/go-pkgz/repeater/v2"
 
 	"github.com/umputun/remark42/backend/app/rest"
+	"github.com/umputun/remark42/backend/app/safehttp"
 	"github.com/umputun/remark42/backend/app/store/image"
 )
 
@@ -153,7 +153,7 @@ func (p Image) downloadImage(ctx context.Context, imgURL string) ([]byte, error)
 
 	transport := p.Transport
 	if transport == nil {
-		transport = ssrfSafeTransport()
+		transport = safehttp.Transport()
 	}
 	client := http.Client{
 		Timeout:   30 * time.Second,
@@ -163,11 +163,11 @@ func (p Image) downloadImage(ctx context.Context, imgURL string) ([]byte, error)
 	var resp *http.Response
 	err := repeater.NewFixed(5, time.Second).Do(ctx, func() error {
 		var e error
-		req, e := http.NewRequest("GET", imgURL, http.NoBody)
+		req, e := http.NewRequest("GET", imgURL, http.NoBody) //nolint:gosec // SSRF mitigated by safehttp.Transport assigned above
 		if e != nil {
 			return fmt.Errorf("failed to make request for %s: %w", imgURL, e)
 		}
-		resp, e = client.Do(req.WithContext(ctx)) //nolint:bodyclose,gosec // body closed in defer; SSRF mitigated by ssrfSafeTransport
+		resp, e = client.Do(req.WithContext(ctx)) //nolint:bodyclose,gosec // body closed in defer; SSRF mitigated by safehttp.Transport
 		return e
 	})
 	if err != nil {
@@ -197,73 +197,4 @@ func (p Image) downloadImage(ctx context.Context, imgURL string) ([]byte, error)
 		return nil, fmt.Errorf("image is too large")
 	}
 	return imgData, nil
-}
-
-// ssrfSafeTransport returns an http.Transport with a dialer that blocks connections to private IP addresses.
-// it resolves the host, validates all IPs, then dials using the resolved IP to prevent DNS rebinding attacks.
-// tries each resolved IP in order to handle dual-stack hosts where the first IP may be unreachable.
-func ssrfSafeTransport() *http.Transport {
-	dialer := &net.Dialer{Timeout: 30 * time.Second}
-	return &http.Transport{
-		DialContext: func(ctx context.Context, network, addr string) (net.Conn, error) {
-			host, port, err := net.SplitHostPort(addr)
-			if err != nil {
-				return nil, fmt.Errorf("invalid address %s: %w", addr, err)
-			}
-
-			// resolve the host to IP addresses
-			ips, err := net.DefaultResolver.LookupIPAddr(ctx, host)
-			if err != nil {
-				return nil, fmt.Errorf("can't resolve host %s: %w", host, err)
-			}
-			if len(ips) == 0 {
-				return nil, fmt.Errorf("no IP addresses resolved for host %s", host)
-			}
-
-			for _, ip := range ips {
-				if isPrivateIP(ip.IP) {
-					return nil, fmt.Errorf("access to private address is not allowed")
-				}
-			}
-
-			// try each resolved IP to handle dual-stack hosts where some IPs may be unreachable
-			var lastErr error
-			for _, ip := range ips {
-				conn, dialErr := dialer.DialContext(ctx, network, net.JoinHostPort(ip.String(), port))
-				if dialErr == nil {
-					return conn, nil
-				}
-				lastErr = dialErr
-			}
-			return nil, fmt.Errorf("can't connect to %s: %w", host, lastErr)
-		},
-	}
-}
-
-// privateCIDRs holds pre-parsed private/reserved CIDR blocks for SSRF protection.
-var privateCIDRs = func() []*net.IPNet {
-	cidrs := []string{
-		"10.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16",
-		"100.64.0.0/10", "127.0.0.0/8", "169.254.0.0/16",
-		"::1/128", "fc00::/7", "fe80::/10",
-	}
-	blocks := make([]*net.IPNet, 0, len(cidrs))
-	for _, cidr := range cidrs {
-		_, block, _ := net.ParseCIDR(cidr)
-		blocks = append(blocks, block)
-	}
-	return blocks
-}()
-
-// isPrivateIP checks if the given IP belongs to a private/reserved range.
-func isPrivateIP(ip net.IP) bool {
-	if ip.IsUnspecified() {
-		return true
-	}
-	for _, block := range privateCIDRs {
-		if block.Contains(ip) {
-			return true
-		}
-	}
-	return false
 }
